undeprecate MDP
