+++ Software/README.md
@@ -1,6 +1,7 @@
 # SPARTAN-Flight-Software
 
-A sensor driver library and main flight loop implemented by the Bruinspace Project Reach Software team. Based on Intel's mraa library for low level i2c communications. 
+A sensor driver library and main flight loop implemented by the Bruinspace Project Reach Software team. Based on
+Intel's [MRAA library](https://github.com/eclipse/mraa) for low level I2C communications.
 
 Table of contents
 =================
@@ -18,13 +19,18 @@ If you're lazy:
 ```shell
 ./build.sh
 ```
-If you like typing stuff:
 
+If you like typing stuff:
 ```shell
 cmake -B _build
 cd _build
 cmake --build .
 ./main/flight_software
+```
+
+Note that after running either command once, for all future builds you only have to run
+```shell
+cmake --build .
 ```
 
 # Sensor documentations
